fix(EPv2): make world_size validation dynamic based on WAVE and schedule
The hardcoded `world_size > 64` check rejected valid configs on wave64
(e.g. EP128 with 2-warp combine) and was too permissive on wave32 with
low warp counts.  Extract _detect_wave_size into wave_config.py (shared
by dispatch_combine_op and intranode_kernels) and replace the static
bound with two checks:

1. Early guard: world_size <= 16 * WAVE (max combine warps in any config)
2. Precise guard after _resolve_geometry: world_size <= min(comb_warp) * WAVE
   across all schedule buckets
