+++ python/mori/ops/dispatch_combine_v2/dispatch_combine_op.py
@@ -50,6 +50,7 @@ from typing import Callable
 import torch
 
 from mori.tensor_utils import from_gpu_ptr
+from .wave_config import WAVE
 
 # Where each backend lives. Imported lazily, on selection only.
 _BACKEND_MODULES = {"flydsl": "flydsl_backend", "hip": "hip_backend"}
@@ -128,14 +129,17 @@ class EpDispatchCombineConfig:
             )
         if self.quant_type != "none":
             self.combine_mode = "scatter"
-        # The combine cross-device barrier uses `tid < npes` to poll per-peer xdb
-        # slots, so npes must fit within a single block (blockDim = warp * WAVE).
-        # The minimum warp count across all schedule variants bounds this; the
-        # hard cap is the maximum wavefront (64) since even a 1-warp block on
-        # wave64 covers that many peers.
-        if self.world_size > 64:
+        # The combine cross-device barrier uses `tid < npes` to poll per-peer
+        # xdb slots, so npes must fit within a single block (blockDim =
+        # combine_warp * WAVE).  The exact bound depends on the resolved
+        # schedule (checked at the end of _resolve_geometry); this early guard
+        # catches obviously invalid values before the tuning lookup runs.
+        # 16 = max combine_warp_num_per_block across all tuning configs
+        # (MI355X and gfx1250 both peak at 16 warps).
+        if self.world_size > 16 * WAVE:
             raise ValueError(
-                f"intranode op supports world_size <= 64, got {self.world_size}"
+                f"intranode op supports world_size <= {16 * WAVE} "
+                f"(16 warps × {WAVE}-wide wave), got {self.world_size}"
             )
         # Token copy moves whole 16 B (vec4) chunks; a non-16 B-aligned per-token
         # size would over-read/write a few dwords past the token.
@@ -235,6 +239,22 @@ class EpDispatchCombineConfig:
                 self.warp_num_per_block = 16
             if self.combine_warp_num_per_block is None:
                 self.combine_warp_num_per_block = 4
+
+        # Precise world_size check against the resolved geometry.  The combine
+        # xdb barrier polls with `tid < npes`, so every schedule bucket must
+        # have blockDim (= comb_warp * WAVE) >= world_size.
+        if self.schedule:
+            min_comb_warp = min(bucket[4] for bucket in self.schedule)
+        else:
+            min_comb_warp = self.combine_warp_num_per_block
+        max_peers = min_comb_warp * WAVE
+        if self.world_size > max_peers:
+            raise ValueError(
+                f"world_size ({self.world_size}) exceeds the smallest combine "
+                f"blockDim in the schedule ({min_comb_warp} warps × {WAVE}-wide "
+                f"wave = {max_peers} threads); the `tid < npes` barrier requires "
+                f"world_size <= blockDim"
+            )
 
     @property
     def is_scatter(self):
